-fix loading issue -refactor the code -fix infinity scroll -search issue fix -add confirmation before delete

+++ src/components/blogs/BlogAction.jsx
@@ -1,21 +1,36 @@
 import PropTypes from "prop-types";
-import { useNavigate } from "react-router-dom";
+import { useLocation, useNavigate } from "react-router-dom";
+import { toast } from "react-toastify";
 import deleteIcon from "../../assets/icons/delete.svg";
 import editIcon from "../../assets/icons/edit.svg";
+import useBlogs from "../../hooks/useBlogs";
 import useProfile from "../../hooks/useProfile";
 
 const BlogAction = ({ blog }) => {
+  const location = useLocation();
   const { deleteBlog } = useProfile();
+  const { removeBlog } = useBlogs();
   const navigation = useNavigate();
-
   const handleEdit = (e) => {
     e.stopPropagation();
+
     navigation(`/blogEdit/${blog.id}`);
   };
 
-  const handleDelete = (e) => {
+  const handleDelete = async (e) => {
+    let msg;
     e.stopPropagation();
-    deleteBlog(blog.id);
+    if (window.confirm("Are you sure you want to Delete this blog?")) {
+      if (location.pathname === "/profile/me") {
+        msg = await deleteBlog(blog.id);
+      } else {
+        msg = await removeBlog(blog.id);
+      }
+    }
+    toast.success(msg, {
+      position: "top-right",
+      autoClose: 2000,
+    });
   };
 
   return (
